refactor(utils): restructure and rename celestial utilities and related imports
Renamed package `utils` to `util` for `CelestialObjectsUtils` to maintain consistency.
Updated imports across the project to reflect the new package structure.
Moved `StarDataRepositoryTest` to align with the updated `map/stars` directory structure.
Removed unused field-of-view constants from `MapViewModel.kt`.
Adjusted formatting and fixed imports in `MapRenderer`, `Moon`, `PlanetsRepository`, and `StarsLoader`.
Cleaned up redundant dependencies and ensured consistency in utility usage.

This refactor improves code organization, readability, and adheres to updated package naming conventions.
