+++ app/src/main/java/com/github/lookupgroup27/lookup/model/map/MapRenderer.kt
@@ -1,17 +1,17 @@
 package com.github.lookupgroup27.lookup.model.map
 
-import PlanetsRepository
 import android.content.Context
 import android.opengl.GLES20
 import android.opengl.GLSurfaceView
 import android.util.Log
 import com.github.lookupgroup27.lookup.R
-import com.github.lookupgroup27.lookup.model.loader.StarsLoader
+import com.github.lookupgroup27.lookup.model.map.planets.PlanetsRepository
 import com.github.lookupgroup27.lookup.model.map.renderables.Planet
 import com.github.lookupgroup27.lookup.model.map.renderables.Star
 import com.github.lookupgroup27.lookup.model.map.renderables.utils.RayUtils.calculateRay
 import com.github.lookupgroup27.lookup.model.map.skybox.SkyBox
 import com.github.lookupgroup27.lookup.model.map.stars.StarDataRepository
+import com.github.lookupgroup27.lookup.model.map.stars.StarsLoader
 import com.github.lookupgroup27.lookup.util.opengl.TextureManager
 import javax.microedition.khronos.egl.EGLConfig
 import javax.microedition.khronos.opengles.GL10
