+++ app/src/main/java/com/github/lookupgroup27/lookup/model/map/stars/StarsLoader.kt
@@ -1,8 +1,7 @@
-package com.github.lookupgroup27.lookup.model.loader
+package com.github.lookupgroup27.lookup.model.map.stars
 
 import android.content.Context
 import com.github.lookupgroup27.lookup.model.map.renderables.Star
-import com.github.lookupgroup27.lookup.model.map.stars.StarDataRepository
 
 /**
  * Converts star data into renderable objects for OpenGL rendering.
